add more tests to performance testing (#7)
* add more graphql api test for call,instruments,pageContent,proposals,questionary,settings and users

* update support files

* update jenkins

* add user calls and userhome browser tests

* improve user setup, make login more consistent and improve test naming

* make user starting id configurable and include it in jenkings file

* remove unnecessary check

* remove steps from template mutations

* update readme file

+++ src/support/userDataSource.ts
@@ -1,29 +1,50 @@
-import { getDatabaseClient } from './database';
+import { getDatabaseClient, sqlQuery } from './database';
 import { numberGenerator, randomUUIDv4 } from '../utils/helperFunctions';
-import {
-  UserLogin,
-  DatabaseClientConnector,
-  DatabaseClient,
-} from '../utils/sharedType';
+import { DatabaseClientConnector, UserLogin } from '../utils/sharedType';
 
 export class UserDataSource {
-  db: DatabaseClient;
-  constructor(username: string, password: string, connectionString: string) {
-    this.db = getDatabaseClient(
-      DatabaseClientConnector.ORACLE,
-      username,
-      password,
-      connectionString
-    );
-  }
+  constructor(
+    private username: string,
+    private password: string,
+    private connectionString: string
+  ) {}
 
-  async getUser(userId: number): Promise<UserLogin> {
+  async checkUserExist(userId: number): Promise<boolean> {
+    let userList = [];
     if (!userId) {
       throw new Error(`UserId is null or undefined: ${userId}`);
     }
-    const transaction = this.db.begin();
+    const db = getDatabaseClient(
+      DatabaseClientConnector.ORACLE,
+      this.username,
+      this.password,
+      this.connectionString
+    );
     try {
-      const uuid = randomUUIDv4();
+      const query = `SELECT USER_NUMBER FROM person WHERE rid = ${userId.toString()}`;
+      userList = sqlQuery(db, query);
+    } catch (error) {
+      throw new Error(`Fail to query user details userId: ${userId}`);
+    } finally {
+      db.close();
+    }
+
+    return userList.length > 0;
+  }
+
+  async createUser(userId: number): Promise<UserLogin> {
+    if (!userId) {
+      throw new Error(`UserId is null or undefined: ${userId}`);
+    }
+    const db = getDatabaseClient(
+      DatabaseClientConnector.ORACLE,
+      this.username,
+      this.password,
+      this.connectionString
+    );
+    const transaction = db.begin();
+    const uuid = randomUUIDv4();
+    try {
       transaction.exec(
         `INSERT INTO address (rid, country, from_date)
            VALUES (:userId, 'United Kingdom', CURRENT_DATE)`,
@@ -35,7 +56,11 @@ export class UserDataSource {
            VALUES (:userId, :userId, :userId, 'Test org', 'Test dept', CURRENT_DATE)`,
         userId
       );
-
+      transaction.exec(
+        `INSERT INTO privacy (privacy_id, searchable)
+             VALUES (:userId,'Yes')`,
+        userId
+      );
       transaction.exec(
         `INSERT INTO person (
                          rid, user_number,
@@ -43,7 +68,7 @@ export class UserDataSource {
                          email, work_phone,
                          establishment_id, org_correspondence_id, home_correspondence_id,
                          from_date, dpa, verified,
-                         isis_salt, sha2, lastpwdreset
+                         isis_salt, sha2, lastpwdreset,privacy_id
                        )
            VALUES (
                     :userId, :userId,
@@ -51,7 +76,7 @@ export class UserDataSource {
                     'BISAPPSSINK' || :userId || '@stfc.ac.uk', '012345123456',
                     :userId, :userId, :userId,
                     CURRENT_DATE, 'TRUE', 'Yes',
-                    'X', 'X', CURRENT_DATE
+                    'X', 'X', CURRENT_DATE,:userId
                   )`,
         userId
       );
@@ -63,11 +88,6 @@ export class UserDataSource {
       );
 
       transaction.commit();
-
-      return {
-        userId,
-        sessionId: uuid,
-      };
     } catch (error) {
       console.error(
         `Error during creating user credentials for userId: ${userId}:`,
@@ -77,18 +97,36 @@ export class UserDataSource {
         transaction.rollback();
       }
       throw new Error(`Fail to create user credentials userId: ${userId}`);
+    } finally {
+      db.close();
     }
+
+    return {
+      userId,
+      sessionId: uuid,
+      email: `BISAPPSSINK${userId}@stfc.ac.uk`,
+    };
   }
 
   async deleteUser(userId: number) {
-    const transaction = this.db.begin();
+    const db = getDatabaseClient(
+      DatabaseClientConnector.ORACLE,
+      this.username,
+      this.password,
+      this.connectionString
+    );
+    const transaction = db.begin();
     try {
       transaction.exec(
         `DELETE FROM login
          WHERE user_id =:userId`,
         userId
       );
-
+      transaction.exec(
+        `DELETE FROM privacy
+          WHERE privacy_id =:userId`,
+        userId
+      );
       transaction.exec(
         `DELETE FROM person
           WHERE user_number =:userId`,
@@ -118,18 +156,26 @@ export class UserDataSource {
         transaction.rollback();
       }
       throw new Error(`Fail to delete user credentials userId: ${userId}`);
+    } finally {
+      db.close();
     }
   }
   async getUsersBetween(
     firstUserId: number,
     lastUserId: number
   ): Promise<UserLogin[]> {
-    await this.deleteUsersBetween(firstUserId, lastUserId);
     const userIdGenerator = numberGenerator(firstUserId);
     const userPromises: Promise<UserLogin>[] = [];
     let userId;
     while ((userId = userIdGenerator.next().value) && userId < lastUserId) {
-      userPromises.push(this.getUser(userId));
+      const userExists = await this.checkUserExist(userId);
+      if (!userExists) {
+        userPromises.push(this.createUser(userId));
+      } else {
+        console.error(
+          `Error during creating user credentials for userId: ${userId} already exists`
+        );
+      }
     }
     const users = await Promise.all(userPromises);
 
@@ -142,34 +188,40 @@ export class UserDataSource {
     if (!lastUserId) {
       throw new Error('Last userId not defined');
     }
-    const transaction = this.db.begin();
+    const db = getDatabaseClient(
+      DatabaseClientConnector.ORACLE,
+      this.username,
+      this.password,
+      this.connectionString
+    );
+    const transaction = db.begin();
     try {
+      transaction.exec(
+        `DELETE FROM person
+          WHERE user_number BETWEEN :firstUserId AND :lastUserId`,
+        firstUserId,
+        lastUserId
+      );
+      transaction.exec(
+        `DELETE FROM privacy
+          WHERE privacy_id BETWEEN :firstUserId AND :lastUserId`,
+        firstUserId,
+        lastUserId
+      );
       transaction.exec(
         `DELETE FROM login
          WHERE user_id BETWEEN :firstUserId AND :lastUserId`,
         firstUserId,
         lastUserId
       );
-
       transaction.exec(
-        `DELETE FROM person
-          WHERE user_number BETWEEN :firstUserId AND :lastUserId
-                OR rid BETWEEN :firstUserId AND :lastUserId`,
+        `DELETE FROM address WHERE rid BETWEEN :firstUserId AND :lastUserId`,
         firstUserId,
         lastUserId
       );
       transaction.exec(
         `DELETE FROM establishment
-          WHERE establishment_id BETWEEN :firstUserId AND :lastUserId
-                OR rid BETWEEN :firstUserId AND :lastUserId`,
-        firstUserId,
-        lastUserId
-      );
-
-      transaction.exec(
-        `DELETE FROM address
-          WHERE postal_address_id BETWEEN :firstUserId AND :lastUserId
-                OR rid BETWEEN :firstUserId AND :lastUserId`,
+          WHERE rid BETWEEN :firstUserId AND :lastUserId`,
         firstUserId,
         lastUserId
       );
@@ -186,6 +238,8 @@ export class UserDataSource {
       throw new Error(
         `Fail to delete user credentials for users ids between: ${firstUserId} and ${lastUserId}`
       );
+    } finally {
+      db.close();
     }
   }
 }
